--删除关于动态代码

+++ src/com/meijialife/simi/ui/calendar/ChinaDate.java
@@ -313,7 +313,7 @@ public final static String getChinaDate(int day) {
 		return "二十";
 	if (day == 30)
 		return "三十";
-	int two = (int) ((day) / 10);
+	int two = (day) / 10;
 	if (two == 0)
 		a = "初";
 	if (two == 1)
@@ -322,7 +322,7 @@ public final static String getChinaDate(int day) {
 		a = "廿";
 	if (two == 3)
 		a = "三";
-	int one = (int) (day % 10);
+	int one = day % 10;
 	switch (one) {
 	case 1:
 		a += "一";
